backend/requirements.txt
# Web framework
fastapi==0.111.1
uvicorn[standard]==0.30.3
python-multipart==0.0.9

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.7.4
pydantic-settings==2.3.4

# Database
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
alembic==1.13.2
geoalchemy2==0.15.2
psycopg2-binary==2.9.9

# Task queue
celery==5.4.0
redis==5.0.7

# File storage
boto3==1.34.144
minio==7.2.7

# 3D model processing
trimesh==4.4.0
numpy==1.26.4
scipy==1.13.1
pygltflib==1.16.2

# IFC processing
ifcopenshell==0.8.4

# Solar simulation (Ladybug Tools ecosystem)
ladybug-core>=0.44.0
ladybug-geometry>=0.31.0
honeybee-core>=0.56.0
honeybee-radiance>=0.66.0
lbt-recipes>=0.23.0

# Report generation
reportlab==4.2.2
pillow==10.4.0

# Utilities
httpx==0.27.0
anyio==4.4.0
python-dotenv==1.0.1
structlog==24.2.0
